Wait 33 ms for the sink to return back the buffer if our queue is empty
This is better than returning an error. However we cannot wait forever otherwise
we deadlock.
Qualcomm camera HAL queues the preview buffer and then tries to dequeue a buffer
immediately. GStreamer sink always keeps a reference to the buffer so dequeue
will always block waiting for the buffer to be returned from the sink which will never
happen and we will simply deadlock.
It seems that Qualcomm HAL is tolerant to the error we return here. To be tested
with camera HAL from other vendors.