Don't send a new segment after flushing viewfinder branch.
If we send a NEWSEGMENT event and then push the first buffer downstream.
What happens is that the sink blocks waiting for clock synchronization.
The real issue is the base_time of our element will be different
than the base_time the sink element. This discrepancy remains until the sink receives
the first buffer after the NEWSEGMENT, prerolls and gets to playing state.
Then out base_time will be updated.
This will stall the pipeline for a while until the update happens.