Don't flush video pipeline before sending EOS while stopping video recording.
When we send FLUSH_STOP, basesink will reset its segment information and set segment
format to GST_FORMAT_UNDEFINED.

When we send an EOS afterwards, mp4mux (qtmux) will send a seek request to filesink
in order to finalize the mp4 file. We will get an assertion because the seek segment format
is in BYTES while the sink format is UNDEFINED.

OTOH flushing the video branch was introduced as a hack to let filesink release its
last buffer otherwise it will contain garbage as Android HAL releases the memory
immediately when we stop video recording.

Since our primary use case is camerabin2 which ensures an encoder (and muxer) will always
be available before filesink then we happily drop this unneeded hack.