We cannot try to acquire auto focus in all modes. Account for that.
We can only try to acquire auto focus if we are in auto or maco modes. If we receive
a request while we are in other modes then we need to stop CAF first before we try to
acquire focus. We will make sure to restore the previous mode back when we stop auto focus.

When we are in any of the CAF modes we switch it to AUTO also because otherwise Android HAL
will not really try to auto focus. It will just send back the status of CAF.