[gstreamer] make sure we send the new segment event before the first buffer. Fixes JB#8765
We were previously checking the status of video recording before waiting
for a buffer to be available and then pushing the buffer we have.

By doing that, we were not accounting for status change that happens between
the loop function sleeping and waking up. which lead to not sending new segment
before the first buffer.

Now we check and see what we need to do after we wake up.