First attempt at a non-flushing pipeline.
Seems flushing is causing issues with GStreamer wrt base_time being changed
after we push the first buffer immediately following a new segment event.

This change in base_time is a result of the sink going back to PLAYING state and is
causing stalls in the pipeline.

What we try to do now is to simply remove any buffers which are not held by us
from our list of buffers. When we need to restart preview again we will simply
allocate new buffers thus we maintain a steady smooth pipeline without flushing.