[gstreamer] Don't reset the focus mode when auto focus is being requested. Fixes JB#10042
When CAF is running, autofocus request will simply return the status of CAF.
It's then up to the application to decide if that's enough or to switch to
the focus mode and run a full autofocus cycle.

We also send a warning if the current focus mode does not allow auto focus to run
but we obey the application.