[gstreamer] Don't free any buffer held by the application. Fixes JB#11053
Issue is we cannot assume that any buffer not in app queue is owned by us.
Buffers pushed downstream are not in app queue yet they are not owned by us anymore.

We simply set and unset a GstBuffer flag depending on whether the buffer has been pushed
downstream or not. This helps us to identify those buffers in order not to unref them.