[droidcamsrc] Don't assume HAL will return back buffers before capturing images.
On mako, HAL likes to keep all the buffers and returns them later on.
We cannot simply toss the buffers otherwise we will crash when HAL cancels them.

The proper way is to use refcounting which also is what HAL expects from the
Android documentation: HAL has a reference between dequeue and either enqueue or cancel).